services/python/requirements.txt
fastapi
uvicorn
python-dotenv
structlog
pydantic
pydantic-settings
pytest
httpx
